+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+### Added
+- Added new `PrivateJoinOnKey` transformation that works with `AddRemoveKeys`
+
 ## 0.5.1 - 2022-11-03
 ### Fixed
 - Domains and metrics make copies of mutable constructor arguments and return copies of mutable properties.